Changes in basicCnnScripy.py
Contains some readme links + a toy code (bnnCnnScript.py) that takes CIFAR10 data set and trains a cnn on it, and than tests the network on the tests set.  I have never ran it and I assume it can take alot of time, I added it just so we hace an example of how the flow should look like.